Removed SBT usage as can't use the external scalac compiler (the new build process is based on https://github.com/jodersky/scala-makefile).
Removed Maven usage as it doesn't use the external javac compiler.

Replaced Stack with Cabal as the former doesn't use the external ghc compiler.

Removed Leiningen usage as it doesn't use the installed Clojure.

Removed unused flag -opt:l:inline for Scala (it works only in if -opt-inline-from is used, and adding the latter actually decreased performance for some tests).