Fix undefined behavior in Zig tests
With current build setting (ReleaseFast), the notify function will
ignore error when there's no server set up. However, it's relying on
undefined behavior to achieve that, whereas with different build
setting it would panic, which is not a desired behavior.

This change makes notify function safely ignore error when connecting
to the server, regardless of build settings, making sure that the test
will continue its execution.

Signed-off-by: Radek Szymanski <radszy@pm.me>