Simplified multiplication of the matrices
The version that I am modifying created a transpose `c` of matrix `b`, which is completely unnecessary. This meant wasted memory and wasted CPU cycles, amounting to a program slower than it should have been for such an elementary task (in particular, it was slower than almost half of the other benchmarked programs written in the other languages). The version that I am proposing is functionally identical, but uses less memory and is faster by dropping the needless intermediate allocations for the transpose.