Temporary fix for delayed server notification on quiet mode
For some reason, in quiet mode, the `notify $ "Idris"` line is delayed
till the end, after the execution is complete, causing a reported 0s
time. Adding `pure ()` at the end mysteriously forces it to be run at
the expected time. This is possibly a compiler bug to be fixed later.